extract get_in fn to fpr.serialize_util
